Add list of other projects in the same space.
I have always liked the community vibe when projects list other projects in the same space. It is a way of saying this idea is important/interesting enough that other people are also working on solutions. We like our solution but we acknowledge there are others out there doing great work also. 👍😀