refactor: simplify ssb.links
rewrite links in a more functional style.
one breaking change - if a source(or source type) was not provided,
previously it defaults to the message as the source. I don't think
this makes sense. you can get the message easily - from the .key
property (and have links return you the whole message with values: true)
Making source always be a feed is also more consistent with the query
input. it doesn't really make sense to query the source as a message
anyway.

The message _is_ an edge, not a vertice.