Use pull-cat to avoid re-broadcasting sequence
Previously the sequence number was being sent to *all listeners*
whenever a new sequence stream was started. This was helpful because it
ensured that listeners would always receive the latest sequence number
when they subscribed to the stream, but bad because it propagated the
current sequence number to all listeners.

Instead, we can concatenate the current sequence number with the future
stream, so that listeners will receive the current + future sequence
numbers without spamming listneres with information they already have.