Replace links view  with ssb-links
I've been reading about how ssbDb.links was meant to be replaced with
ssb-links but it hasn't happened, so I wanted to try it. Since ssb-links
doesn't have its own tests we have to rely on ssb-db's tests, which pass
with one exception: self-links! It looks like ssb-links will generate an
item when you mention yourself but ssb-db's links() doesn't, which is
why the tests are failing on this commit.

What's the right move here? I'd love to wrap up this intention.