Add setImmediate() to hack test/close.js
Problem: The sweet new work on unboxers and queues has introduced a
small regression where you can't turn off the server and turn it back on
again in the same tick without issues. It's unclera why this is
happening, but I'm happy to trade off ergonomics and private groups for
a small quirk when you flick the light switch on and off again too fast.

Solution: Use `setImmediate()` to make tests fail. This commit assumes
that a note will also go into the readme explaining this bug. SHIP IT.