Add rebuild method
Problem: There's no way to rebuild the database from SSB-DB, which
is necessary when you get a new unboxer key or want to resolve some
weird database bug under the hood.

Solution: Add `rebuild()` method to database and add tests to ensure
that it has the expected behavior. Currently it does **not**, and this
suggests that FlumeDB's rebuild code isn't actually doing any rebuilding
under the hood.