Add full tests for index.js, not create.js
Problem: Running tests that import from `create.js` means that we're not
actually testing the module that we're exporting to applications that
depend on SSB-DB. This means that our test have different behavior in
development versus production, and makes it easier for subtle bugs to
split in.

Solution: Test against `index.js`, not `create.js`, and fix all of the
problems that occur when you test against the full module instead of the
previous subset.