Fix ghost messages by disabling unbox cache
Problem: Some messages aren't getting unboxed correctly.

Solution: Disable the unbox cache, which seems to be one possible way to
solve this problem. I'm going to continue investigating alternatives so
that we don't need to lose the cache that @mixmix built out, but this
might be a Good Enough solution for now.