Fix originalValue() shared mutable state
Problem: The `originalValue()` function is mutating the message
directly, which apparently causes those changes to be mirrored in every
other reference to the message. I've only been able to reproduce this
problem in stream-based methods, and have written a test to prove it.

Solution: Refactor `originalData()` and `originalValue()` to make copies
of the data before passing to `mutateValue()` for the actual re-boxing
behavior.