Fix known close() bug
Problem: There's a problem where `close()` calls back before it's
actually closed, which means that people have to add annoying
`setImmediate()` boilerplate to their tests. I also have a hunch that
this is causing all of the Windows test failures that are driving me up
the wall.

Solution: Do the `setImmediate()` before `close()`. I'm embarassed that
it took me so long to come up with this as a solution.