Fix pass-by-reference from cache
Problem: HashLRU passes references, not values, so when we use the cache
we're accidentally creating a scenario where one variable being mutated
can change the contents of the results of another query.

Solution: Use `lodash.cloneDeep()` on the results from the HashLRU cache
to ensure that we're always passing by value rather than by reference.