Pass feed state to boxer
Problem: Some boxers, like SSB-Tribe, depend on the feed state. We've
tried using hacks to query the database before publishing a message, but
we really just need to be passed the feed state (according to the
validator) each time the boxer is asked to box a message.

Solution: Pass the relevant feed state from the validator to the boxer.